move rdf dependent tests to datadriven
